requirements.txt
streamlit==1.45.1
google-generativeai==0.8.5
python-docx==1.1.2
PyPDF2==3.0.1
python-dotenv==1.1.0
